Reorganize code
This moves all executables to cmd/ and library code to the top-level.

Resolves #5
