+++ cmd/parallel-exec/README.md
