+++ cmd/update-license/main.go
