+++ cmd/update-license/main_test.go
