+++ examples/parallel-exec/bin/simple.sh
