+++ parallel/event.go
