+++ parallel/exec_cmd.go
