+++ parallel/parallel.go
