+++ parallel/runner.go
