+++ parallel/testdata/bin/simple.sh
